Restructure storage component
