+++ src/main/java/seedu/address/MainApp.java
@@ -24,15 +24,15 @@ import seedu.address.model.ReadOnlyPatientBook;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.util.SampleDataUtil;
-import seedu.address.storage.AppointmentBookStorage;
-import seedu.address.storage.JsonAppointmentBookStorage;
-import seedu.address.storage.JsonPatientBookStorage;
 import seedu.address.storage.JsonUserPrefsStorage;
-import seedu.address.storage.PatientBookStorage;
 import seedu.address.storage.Storage;
 import seedu.address.storage.StorageManager;
 import seedu.address.storage.StorageStatsManager;
 import seedu.address.storage.UserPrefsStorage;
+import seedu.address.storage.appointment.AppointmentBookStorage;
+import seedu.address.storage.appointment.JsonAppointmentBookStorage;
+import seedu.address.storage.patient.JsonPatientBookStorage;
+import seedu.address.storage.patient.PatientBookStorage;
 import seedu.address.ui.Ui;
 import seedu.address.ui.UiManager;
