+++ src/main/java/seedu/address/storage/appointment/JsonAppointmentBookStorage.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.appointment;
 
 import static java.util.Objects.requireNonNull;
 
@@ -12,6 +12,9 @@ import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.commons.util.JsonUtil;
 import seedu.address.model.ReadOnlyAppointmentBook;
+import seedu.address.storage.StorageStatsManager;
+import seedu.address.storage.archive.AppointmentArchive;
+import seedu.address.storage.archive.CsvAppointmentArchive;
 
 /**
  * A class to access AppointmentBook data stored as a json file on the hard disk.
