+++ src/main/java/seedu/address/storage/archive/AppointmentArchive.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.archive;
 
 import java.io.IOException;
 import java.nio.file.Path;
