+++ src/main/java/seedu/address/storage/patient/JsonPatientBookStorage.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.patient;
 
 import static java.util.Objects.requireNonNull;
 
@@ -12,6 +12,7 @@ import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.commons.util.JsonUtil;
 import seedu.address.model.ReadOnlyPatientBook;
+import seedu.address.storage.StorageStatsManager;
 
 /**
  * A class to access PatientBook data stored as a json file on the hard disk.
