+++ src/test/java/seedu/address/storage/archive/CsvAdaptedPatientTest.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.archive;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
